requirements.txt
# Dépendances principales du projet Data Science

# Manipulation de données
pandas==2.0.3
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.0
xgboost==2.0.0

# Deep Learning (optionnel)
# tensorflow==2.13.0
# torch==2.0.1

# Visualisation
matplotlib==3.7.2
seaborn==0.12.2

# API Web
Flask==2.3.2
Flask-Cors==4.0.0

# Gestion de fichiers
PyYAML==6.0.1
python-dotenv==1.0.0

# Sérialisation
dill==0.3.7

# Notebooks
jupyter==1.0.0
notebook==7.0.0
ipykernel==6.25.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Qualité du code
black==23.7.0
flake8==6.1.0
pylint==2.17.5

# MLOps (optionnel - décommenter si nécessaire)
# mlflow==2.5.0
# dvc==3.15.0

# Cloud (optionnel - décommenter selon besoin)
# boto3==1.28.17  # AWS
# google-cloud-storage==2.10.0  # GCP
# azure-storage-blob==12.17.0  # Azure

# Base de données (optionnel)
# psycopg2-binary==2.9.7  # PostgreSQL
# pymongo==4.4.1  # MongoDB
# sqlalchemy==2.0.19  # ORM

# Monitoring et logging
# sentry-sdk==1.29.2

# Utilitaires
tqdm==4.66.1
requests==2.31.0

# Installation en mode développement (pour setup.py)
-e .
